+++ tests/properties.rs
@@ -40,6 +40,30 @@ proptest! {
     }
 
     #[test]
+    fn decode_then_encode_canonical(possibly_encoded_string in "[\t\n\r -~]*") {
+        // Iterate over the string until we hit a decoding error. Then, get
+        // the index of the iterator (the index where we hit the error).
+        let mut possibly_encoded_iter = possibly_encoded_string.as_bytes().iter();
+        let _decode_error = tick_encoding::decode_iter(possibly_encoded_iter.by_ref().copied()).find_map(|result| result.err());
+        let first_invalid_index = possibly_encoded_string.len() - possibly_encoded_iter.as_slice().len();
+
+        // Because we may have hit an `UnexpectedEnd` error, we might need to
+        // cut a few more bytes off the end of the string to get a valid
+        // encoding
+        let mut encoded = &possibly_encoded_string[..first_invalid_index];
+        while tick_encoding::decode(encoded.as_bytes()).is_err() {
+            encoded = &encoded[..encoded.len() - 1];
+        }
+
+        let decoded = tick_encoding::decode(encoded.as_bytes()).unwrap();
+        let re_encoded = tick_encoding::encode(&decoded);
+        let re_decoded = tick_encoding::decode(re_encoded.as_bytes()).unwrap();
+
+        assert_eq!(encoded, re_encoded);
+        assert_eq!(decoded, re_decoded);
+    }
+
+    #[test]
     fn encode_unescaped_borrows(unescaped_string in "[\t\n\r -_a-~]*") {
         let encoded = tick_encoding::encode(unescaped_string.as_bytes());
         assert_matches!(encoded, Cow::Borrowed(_));