+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed a bug where the sequence "\`60" would be accepted and decode to "\`". The canonical encoding is "\`\`", so this now returns an error.
+
 ## [v0.1.0] - 2024-01-28
 
 ### Added